Table Version
I've attempted to convert the leader board list into a table rather than a list, so that it better matches the style of other affiliate area tabs, such as the stats area tables.

My php isn't great as you can probably tell, so this is as far as I got. I've got two columns, one for name, the other for the out put of visits/referrals/earnings.

I've lost the number position count that came with the OL tag, but with the right CSS I've got a responsive table.

If anyone can improve, individual columns for 'Visits' 'Referrals' & 'Earnings' would be good.

If they could be sortable, with pagination of all users, that would be cool.